+++ packages/typespec-go/test/azure-http-specs/azure/client-generator-core/accessgroup/relativemodelinoperation_client_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestRelativeModelInOperationClient_discriminator(t *testing.T) {
-	client, err := NewAccessClient("http://localhost:3000", nil)
+	client, err := NewAccessClientWithNoCredential("http://localhost:3000", nil)
 	require.NoError(t, err)
 	resp, err := client.NewAccessRelativeModelInOperationClient().discriminator(context.Background(), "real", nil)
 	require.NoError(t, err)
@@ -23,7 +23,7 @@ func TestRelativeModelInOperationClient_discriminator(t *testing.T) {
 }
 
 func TestRelativeModelInOperationClient_operation(t *testing.T) {
-	client, err := NewAccessClient("http://localhost:3000", nil)
+	client, err := NewAccessClientWithNoCredential("http://localhost:3000", nil)
 	require.NoError(t, err)
 	resp, err := client.NewAccessRelativeModelInOperationClient().operation(context.Background(), "Madge", nil)
 	require.NoError(t, err)